move QueryArguments to ApiConnector
