+++ server/connectors/api-connector-mysql/src/main/scala/com/prisma/api/connector/mysql/database/DataResovler.scala
@@ -1,5 +1,5 @@
 package com.prisma.api.connector.mysql.database
-import com.prisma.api.connector.{DataItem, NodeSelector}
+import com.prisma.api.connector.{DataItem, NodeSelector, QueryArguments}
 import com.prisma.api.connector.mysql.database.Types.DataItemFilterCollection
 import com.prisma.shared.models.IdType.Id
 import com.prisma.shared.models.{Field, Model, Project, Relation}
