+++ server/connectors/api-connector-mysql/src/main/scala/com/prisma/api/connector/mysql/database/Types.scala
@@ -1,10 +1,7 @@
 package com.prisma.api.connector.mysql.database
 
-import com.prisma.api.connector.DataItem
-import com.prisma.api.connector.mysql.database.Types.{DataItemFilterCollection}
-import com.prisma.shared.models.IdType.Id
+import com.prisma.api.connector.mysql.database.Types.DataItemFilterCollection
 import com.prisma.shared.models.{Field, Model, Relation}
-import sangria.relay.Node
 
 object Types {
   type DataItemFilterCollection = Seq[_ >: Seq[Any] <: Any]
@@ -20,14 +17,3 @@ case class FilterElement(key: String,
 case class FilterElementRelation(fromModel: Model, toModel: Model, relation: Relation, filter: DataItemFilterCollection)
 
 case class ScalarListValue(nodeId: String, position: Int, value: Any)
-
-object SortOrder extends Enumeration {
-  type SortOrder = Value
-  val Asc: SortOrder.Value  = Value("asc")
-  val Desc: SortOrder.Value = Value("desc")
-}
-
-case class OrderBy(
-    field: Field,
-    sortOrder: SortOrder.Value
-)
