--- server/connectors/api-connector/src/main/scala/com/prisma/api/connector/DataItemFilterCollection.scala
@@ -1,6 +1,0 @@
-package com.prisma.api.connector
-
-object Types {
-  type DataItemFilterCollection = Seq[_ >: Seq[Any] <: Any]
-  //  type UserData                 = Map[String, Option[Any]]
-}
