+++ server/servers/api/src/main/scala/com/prisma/api/resolver/DeferredTypes.scala
@@ -1,7 +1,6 @@
 package com.prisma.api.resolver
 
-import com.prisma.api.connector.DataItem
-import com.prisma.api.connector.mysql.database.QueryArguments
+import com.prisma.api.connector.{DataItem, QueryArguments}
 import com.prisma.shared.models.{Field, Model}
 import sangria.execution.deferred.Deferred
