+++ server/servers/api/src/main/scala/com/prisma/api/resolver/DeferredUtils.scala
@@ -1,6 +1,6 @@
 package com.prisma.api.resolver
 
-import com.prisma.api.connector.mysql.database.QueryArguments
+import com.prisma.api.connector.QueryArguments
 import com.prisma.api.resolver.DeferredTypes._
 import com.prisma.shared.models.IdType.Id
 import com.prisma.shared.models.{Field, Model}
