+++ server/servers/api/src/main/scala/com/prisma/api/resolver/IdBasedConnection.scala
@@ -1,5 +1,6 @@
 package com.prisma.api.resolver
 
+import com.prisma.api.connector.QueryArguments
 import com.prisma.api.connector.mysql.database._
 import com.prisma.shared.models
 import sangria.schema._
